requirements.txt
pandas
numpy
geopandas
networkx
shapely
pycountry
scikit-learn
requests
matplotlib
jupyter
psycopg2-binary